+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# zuora_client CHANGELOG
+zuora_client CHANGELOG
+======================
 
 2.0.1-orm
 ---------